+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Stop using typolink for external links (#1123)
 - Drop the `TemplateHelper` dependency from the legacy models (#1117)
 - Use the Extbase localization features in the legacy models (#1116)
-- Switch to the new configuration classes in more places (#1112, #1114)
+- Switch to the new configuration classes in more places (#1112, #1114, #1126)
 - Stop calling `initTemplate()` in the link builder (#1098)
 - Use PHP 7.2 features (#1095)
 - Raise PHPStan to level 6 (#1093)